+++ src/components/user/index.ts
@@ -1,0 +1,4 @@
+export { default as ProfileButton } from './ProfileButton';
+export { default as UserAvatar } from './UserAvatar';
+export { default as UserInfoSection } from './UserInfoSection';
+export { default as UserProfileSection } from './UserProfileSection';
